+++ src/main/java/com/se/pickple_api_server/v1/account/application/dto/AccountReadDto.java
@@ -3,8 +3,8 @@ package com.se.pickple_api_server.v1.account.application.dto;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.se.pickple_api_server.v1.account.domain.entity.Account;
-import com.se.pickple_api_server.v1.account.domain.entity.AccountType;
-import com.se.pickple_api_server.v1.account.domain.entity.RegisterType;
+import com.se.pickple_api_server.v1.account.domain.type.AccountType;
+import com.se.pickple_api_server.v1.account.domain.type.RegisterType;
 import com.se.pickple_api_server.v1.common.infra.dto.PageRequest;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
