+++ src/main/java/com/se/pickple_api_server/v1/account/application/service/AccountCreateService.java
@@ -4,8 +4,8 @@ package com.se.pickple_api_server.v1.account.application.service;
 import com.se.pickple_api_server.v1.account.application.dto.AccountCreateDto;
 import com.se.pickple_api_server.v1.account.application.error.AccountErrorCode;
 import com.se.pickple_api_server.v1.account.domain.entity.Account;
-import com.se.pickple_api_server.v1.account.domain.entity.AccountType;
-import com.se.pickple_api_server.v1.account.domain.entity.RegisterType;
+import com.se.pickple_api_server.v1.account.domain.type.AccountType;
+import com.se.pickple_api_server.v1.account.domain.type.RegisterType;
 import com.se.pickple_api_server.v1.account.infra.repository.AccountJpaRepository;
 import com.se.pickple_api_server.v1.common.domain.exception.BusinessException;
 import com.se.pickple_api_server.v1.oauth.infra.dto.userinfo.OauthUserInfo;
