+++ src/main/java/com/se/pickple_api_server/v1/account/application/service/AccountUpdateService.java
@@ -3,12 +3,11 @@ package com.se.pickple_api_server.v1.account.application.service;
 import com.se.pickple_api_server.v1.account.application.dto.AccountUpdateDto;
 import com.se.pickple_api_server.v1.account.application.error.AccountErrorCode;
 import com.se.pickple_api_server.v1.account.domain.entity.Account;
-import com.se.pickple_api_server.v1.account.domain.entity.AccountType;
+import com.se.pickple_api_server.v1.account.domain.type.AccountType;
 import com.se.pickple_api_server.v1.account.infra.repository.AccountJpaRepository;
 import com.se.pickple_api_server.v1.common.domain.error.GlobalErrorCode;
 import com.se.pickple_api_server.v1.common.domain.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
