--- src/main/java/com/se/pickple_api_server/v1/account/domain/entity/AccountType.java
@@ -1,5 +1,0 @@
-package com.se.pickple_api_server.v1.account.domain.entity;
-
-public enum AccountType {
-    UNKNOWN, MEMBER, ADMIN
-}
