+++ src/main/java/com/se/pickple_api_server/v1/account/domain/type/AccountType.java
@@ -1,0 +1,5 @@
+package com.se.pickple_api_server.v1.account.domain.type;
+
+public enum AccountType {
+    UNKNOWN, MEMBER, ADMIN
+}
