+++ src/main/java/com/se/pickple_api_server/v1/account/domain/type/RegisterType.java
@@ -1,4 +1,4 @@
-package com.se.pickple_api_server.v1.account.domain.entity;
+package com.se.pickple_api_server.v1.account.domain.type;
 
 public enum RegisterType {
     NAVER, PICKPLE, NORMAL, SE
