+++ src/main/java/com/se/pickple_api_server/v1/apply/domain/entity/Apply.java
@@ -1,11 +1,9 @@
 package com.se.pickple_api_server.v1.apply.domain.entity;
 
-import com.se.pickple_api_server.v1.board.domain.entity.Board;
-import com.se.pickple_api_server.v1.board.domain.entity.RecruitmentBoard;
+import com.se.pickple_api_server.v1.recruitment.domain.entity.RecruitmentBoard;
 import com.se.pickple_api_server.v1.common.domain.entity.BaseEntity;
 import com.se.pickple_api_server.v1.profile.domain.entity.Profile;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
