+++ src/main/java/com/se/pickple_api_server/v1/apply/infra/repository/ApplyJpaRepository.java
@@ -2,8 +2,7 @@ package com.se.pickple_api_server.v1.apply.infra.repository;
 
 import com.se.pickple_api_server.v1.account.domain.entity.Account;
 import com.se.pickple_api_server.v1.apply.domain.entity.Apply;
-import com.se.pickple_api_server.v1.board.domain.entity.Board;
-import com.se.pickple_api_server.v1.board.domain.entity.RecruitmentBoard;
+import com.se.pickple_api_server.v1.recruitment.domain.entity.RecruitmentBoard;
 import com.se.pickple_api_server.v1.profile.domain.entity.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
