+++ src/main/java/com/se/pickple_api_server/v1/board/domain/entity/Board.java
@@ -1,9 +1,9 @@
 package com.se.pickple_api_server.v1.board.domain.entity;
 
+import com.se.pickple_api_server.v1.board.domain.type.BoardType;
 import com.se.pickple_api_server.v1.common.domain.entity.BaseEntity;
 import com.se.pickple_api_server.v1.account.domain.entity.Account;
 import lombok.*;
-import org.hibernate.annotations.Cascade;
 
 import javax.persistence.*;
 import javax.validation.constraints.Size;
