--- src/main/java/com/se/pickple_api_server/v1/board/domain/entity/BoardType.java
@@ -1,5 +1,0 @@
-package com.se.pickple_api_server.v1.board.domain.entity;
-
-public enum BoardType {
-    RECRUITMENT,
-}
