+++ src/main/java/com/se/pickple_api_server/v1/board/domain/type/BoardType.java
@@ -1,0 +1,5 @@
+package com.se.pickple_api_server.v1.board.domain.type;
+
+public enum BoardType {
+    RECRUITMENT,
+}
