+++ src/main/java/com/se/pickple_api_server/v1/bookmark/infra/repository/BookmarkJpaRepository.java
@@ -1,12 +1,10 @@
 package com.se.pickple_api_server.v1.bookmark.infra.repository;
 
 import com.se.pickple_api_server.v1.account.domain.entity.Account;
-import com.se.pickple_api_server.v1.board.domain.entity.Board;
-import com.se.pickple_api_server.v1.board.domain.entity.RecruitmentBoard;
+import com.se.pickple_api_server.v1.recruitment.domain.entity.RecruitmentBoard;
 import com.se.pickple_api_server.v1.bookmark.domain.entity.Bookmark;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.awt.print.Book;
 import java.util.List;
 import java.util.Optional;
