--- src/main/java/com/se/pickple_api_server/v1/oauth/infra/dto/token/NaverOauthTokenResponse.java
@@ -1,37 +1,0 @@
-//package com.se.pickple_api_server.v1.oauth.infra.dto.token;
-//
-//import lombok.AllArgsConstructor;
-//import lombok.Builder;
-//import lombok.Data;
-//import lombok.NoArgsConstructor;
-//
-//@Data
-//@NoArgsConstructor
-//@AllArgsConstructor
-//@Builder
-//public class NaverOauthTokenResponse implements OauthTokenResponse {
-//    String access_token;
-//    String refresh_token;
-//    String token_type;
-//    Long expires_in;
-//
-//    @Override
-//    public String getAccessToken() {
-//        return access_token;
-//    }
-//
-//    @Override
-//    public String getRefreshToken() {
-//        return refresh_token;
-//    }
-//
-//    @Override
-//    public String getTokenType() {
-//        return token_type;
-//    }
-//
-//    @Override
-//    public Long getExpiresIn() {
-//        return expires_in;
-//    }
-//}
