--- src/main/java/com/se/pickple_api_server/v1/oauth/infra/dto/token/OauthTokenResponse.java
@@ -1,8 +1,0 @@
-//package com.se.pickple_api_server.v1.oauth.infra.dto.token;
-//
-//public interface OauthTokenResponse {
-//    String getAccessToken();
-//    String getRefreshToken();
-//    String getTokenType();
-//    Long getExpiresIn();
-//}
