+++ src/main/java/com/se/pickple_api_server/v1/oauth/infra/dto/userinfo/OauthUserInfo.java
@@ -3,6 +3,5 @@ package com.se.pickple_api_server.v1.oauth.infra.dto.userinfo;
 public interface OauthUserInfo {
     String getId();
     String getName();
-    //String getMobile();
     String getEmail();
 }
