+++ src/main/java/com/se/pickple_api_server/v1/recruitment/infra/repository/BoardJpaRepository.java
@@ -1,4 +1,4 @@
-package com.se.pickple_api_server.v1.board.infra.repository;
+package com.se.pickple_api_server.v1.recruitment.infra.repository;
 
 import com.se.pickple_api_server.v1.board.domain.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
