+++ src/main/java/com/se/pickple_api_server/v1/recruitment/infra/repository/RecruitmentBoardJpaRepository.java
@@ -1,7 +1,7 @@
-package com.se.pickple_api_server.v1.board.infra.repository;
+package com.se.pickple_api_server.v1.recruitment.infra.repository;
 
 import com.se.pickple_api_server.v1.account.domain.entity.Account;
-import com.se.pickple_api_server.v1.board.domain.entity.RecruitmentBoard;
+import com.se.pickple_api_server.v1.recruitment.domain.entity.RecruitmentBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
