--- src/main/java/com/se/pickple_api_server/v1/report/domain/entity/ReportResult.java
@@ -1,5 +1,0 @@
-package com.se.pickple_api_server.v1.report.domain.entity;
-
-public enum ReportResult {
-    NONE, POST_DELETED
-}
