--- src/main/java/com/se/pickple_api_server/v1/report/domain/entity/ReportState.java
@@ -1,5 +1,0 @@
-package com.se.pickple_api_server.v1.report.domain.entity;
-
-public enum ReportState {
-    BEFORE, AFTER
-}
