+++ src/main/java/com/se/pickple_api_server/v1/report/domain/type/ReportResult.java
@@ -1,0 +1,5 @@
+package com.se.pickple_api_server.v1.report.domain.type;
+
+public enum ReportResult {
+    NONE, POST_DELETED
+}
