+++ src/main/java/com/se/pickple_api_server/v1/report/domain/type/ReportState.java
@@ -1,0 +1,5 @@
+package com.se.pickple_api_server.v1.report.domain.type;
+
+public enum ReportState {
+    BEFORE, AFTER
+}
