+++ src/main/java/com/se/pickple_api_server/v1/tag/domain/entity/Tag.java
@@ -1,8 +1,6 @@
 package com.se.pickple_api_server.v1.tag.domain.entity;
 
-import com.se.pickple_api_server.v1.board.domain.entity.RecruitmentBoard;
 import com.se.pickple_api_server.v1.common.domain.entity.BaseEntity;
-import com.se.pickple_api_server.v1.profile.domain.entity.Profile;
 import lombok.*;
 
 import javax.persistence.*;
